Move embedding to utilities folder
