+++ frontend/src/components/design-system/style/BackgroundComponent.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { computed, inject } from 'vue';
-import { EmbedKey } from '@/components/EmbedKey';
+import { EmbedKey } from '@/components/utilities/embedding/EmbedKey';
 
 export interface Props {
   isImage?: boolean;
