+++ frontend/src/components/utilities/embedding/EmbedKey.ts
