+++ frontend/src/components/utilities/embedding/EmbedProviderWrapper.vue
@@ -2,8 +2,8 @@
 import { computed, provide } from 'vue';
 import { useRoute } from 'vue-router';
 
-import EmbedProvider from '@/components/EmbedProvider.vue';
-import { EmbedKey } from '@/components/EmbedKey';
+import EmbedProvider from '@/components/utilities/embedding/EmbedProvider.vue';
+import { EmbedKey } from '@/components/utilities/embedding/EmbedKey';
 
 const route = useRoute();
