+++ frontend/src/routes/question/QuestionPage.vue
@@ -21,7 +21,7 @@ import ResponsiveWrapper from '@/components/responsivity/ResponsiveWrapper.vue';
 import SecondaryNavigationBar from '@/components/design-system/navigation/SecondaryNavigationBar.vue';
 import StepWrapper from '@/components/design-system/layout/StepWrapper.vue';
 
-import EmbedWrapper from '@/components/responsivity/EmbedWrapper.vue';
+import EmbedWrapper from '@/components/utilities/embedding/EmbedWrapper.vue';
 import QuestionBottomBar from './QuestionBottomBar.vue';
 import QuestionCard from './QuestionCard.vue';
 import BackgroundComponent from '../../components/design-system/style/BackgroundComponent.vue';
