+++ frontend/src/routes/recap/RecapPage.vue
@@ -23,7 +23,7 @@ import TitleText from '@/components/design-system/typography/TitleText.vue';
 
 import { vkiLogoPercent } from '@/components/design-system/icons';
 
-import EmbedWrapper from '@/components/responsivity/EmbedWrapper.vue';
+import EmbedWrapper from '@/components/utilities/embedding/EmbedWrapper.vue';
 import ResponsiveWrapper from '@/components/responsivity/ResponsiveWrapper.vue';
 import StickyHeaderLayout from '@/components/layouts/StickyHeaderLayout.vue';
