+++ frontend/src/routes/share/SharePage.vue
@@ -19,7 +19,7 @@ import SecondaryNavigationBar from '@/components/design-system/navigation/Second
 import StackComponent from '@/components/design-system/layout/StackComponent.vue';
 import TitleText from '@/components/design-system/typography/TitleText.vue';
 
-import EmbedWrapper from '@/components/responsivity/EmbedWrapper.vue';
+import EmbedWrapper from '@/components/utilities/embedding/EmbedWrapper.vue';
 import ResponsiveWrapper from '@/components/responsivity/ResponsiveWrapper.vue';
 import StickyHeaderLayout from '@/components/layouts/StickyHeaderLayout.vue';
 import { mdiRepeat } from '@mdi/js';
